news/requirements.txt
pandas>=1.4.2
networkx~=2.6.3 #2.7.0 is not work
ckip-transformers~=0.3.2
# torch>=1.5.0
# transformers>=3.5.0

loguru>=0.6.0
opencc-python-reimplemented~=0.1.6
pytrends>=4.8.0
scipy~=1.7.3
spacy~=3.4.3

Flask~=2.2.2
requests~=2.28.1
typer~=0.7.0

jieba~=0.42.1
tqdm>=4.27
cn2an~=0.5.19
# ckipnlp

# ----crawler------
# beautifulsoup4~=4.10.0
# GitPython~=3.1.29
# lxml~=4.9.1